% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tree.R
\name{prune_tree}
\alias{prune_tree}
\title{Prunes an \code{rpart} tree to have the desired number of nodes.}
\usage{
prune_tree(tree, maxn)
}
\arguments{
\item{tree}{the tree to prune, an \code{rpart} object.}

\item{maxn}{desired number of terminal nodes.}
}
\value{
The pruned tree, an \code{rpart} object.
}
\description{
Prunes an \code{rpart} tree to have the desired number of nodes.
}
