% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surve.R
\name{surve}
\alias{surve}
\title{Defines the evaluation function for a new splitting method of \code{rpart}.}
\usage{
surve(y, wt, parms)
}
\arguments{
\item{y}{the response value as found in the formula that is passed in by \code{rpart}.
  Note that \code{rpart} will normally have removed any observations 
with a missing response.}

\item{wt}{the weight vector from the call, if any.}

\item{parms}{the vector or list (if any) supplied by the user as a
\code{parms} argument to the call.}
}
\value{
See reference.
}
\description{
Defines the evaluation function for a new splitting method of \code{rpart}.
 Not to be called directly by the user.
}
\references{
\url{https://cran.r-project.org/package=rpart/vignettes/usercode.pdf}
}
\seealso{
\code{\link{survs},\link{survi}}
}
