% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survi.R
\name{survi}
\alias{survi}
\title{Defines the initialization function for a new splitting method of \code{rpart}.}
\usage{
survi(y, offset, parms, wt)
}
\arguments{
\item{y}{the response value as found in the formula that is passed in by \code{rpart}.
Note that \code{rpart} will normally
have removed any observations with a missing response.}

\item{offset}{the offset term, if any, found on the right hand side of the 
formula that is passed in by \code{rpart}.}

\item{parms}{the vector or list (if any) supplied by the user as a
\code{parms} argument to the call.}

\item{wt}{the weight vector from the call, if any.}
}
\value{
See reference.
}
\description{
Defines the initialization function for a new splitting method of \code{rpart}.
 Not to be called directly by the user.
}
\references{
\url{https://cran.r-project.org/package=rpart/vignettes/usercode.pdf}
}
\seealso{
\code{\link{survs},\link{surve}}
}
